requirements.txt
python-telegram-bot
typing
nltk